requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
pyyaml>=6.0
aiohttp>=3.9.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
